fix(recipes): render create page route
